// code/frontend/src/enums.tsx
// Enum para Dificuldade
export enum Difficulty {
    EASY = "Easy",
    MEDIUM = "Medium",
    HARD = "Hard",
  }
  
  // Enum para Categoria
  export enum Category {
    APPETIZER = "Appetizer",
    MAIN_COURSE = "Main Course",
    DESSERT = "Dessert",
    BEVERAGE = "Beverage",
  }
  
  // Enum para Tipo de Cozinha
  export enum CuisineType {
    ITALIAN = "italian",
    BRAZILIAN = "Brazilian",
    CHINESE = "Chinese",
    INDIAN = "Indonesian",
    JAPANESE = "Japanese",
    GERMAN = "German",
    THAI = "Thai",
    MEXICAN = "Mexican",
    FRENCH = "French",
    USA = "United States",
  }
  
  // Enum para Tipo de Dieta
  export enum DietType {
    VEGETARIAN = "Vegetarian",
    VEGAN = "Vegan",
    CARNIVORE = "Carnivore",
    LOW_CARB = "Low Carb",
    KETO = "Keto",
    MEDITERRANEAN = "Mediterranean",
    GLUTEN_FREE = "Gluten Free",
    LACTOSE_FREE = "Lactose Free",
  }
  
  // Enum para Unidade de Medida
  export enum UnitType {
    GRAMS = "Grams",
    KILOGRAMS = "Kilograms",
    LITERS = "Liters",
    TABLESPOONS = "Tablespoons",
    CUPS = "Cups",
    TEASPOONS = "Teaspoons",
    UNITS = "Units",
  }
  